+++ app/(tabs)/profile.js
@@ -1,0 +1,118 @@
+import { Button, SafeAreaView, ScrollView, StyleSheet, Text, View } from 'react-native';
+import { useAuth } from '../../contexts/AuthContext';
+
+
+export default function ExploreScreen() {
+const { auth, setAuth } = useAuth();
+
+    const handleLogout = () => {
+    setAuth(null);
+  }
+
+
+  return (
+    <SafeAreaView style={styles.container}>
+      <ScrollView style={styles.scrollView} contentContainerStyle={styles.content}>
+        <View style={styles.header}>
+          <Text style={styles.title}>Perfil</Text>
+          <Text style={styles.subtitle}>
+            Información sobre la aplicación Mixer
+          </Text>
+        </View>
+
+
+        <View style={styles.card}>
+          <Button title="Cerrar sesión" onPress={handleLogout} /> 
+        </View>
+
+
+        <View style={styles.footer}>
+          <Text style={styles.footerText}>
+            Mixer App v1.0 - Creado con React Native y Expo
+          </Text>
+        </View>
+      </ScrollView>
+    </SafeAreaView>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#f8f9fa',
+  },
+  scrollView: {
+    flex: 1,
+  },
+  content: {
+    padding: 20,
+    paddingBottom: 40,
+  },
+  header: {
+    marginBottom: 30,
+    alignItems: 'center',
+  },
+  title: {
+    fontSize: 32,
+    fontWeight: 'bold',
+    color: '#212529',
+    marginBottom: 8,
+  },
+  subtitle: {
+    fontSize: 16,
+    color: '#6c757d',
+    textAlign: 'center',
+  },
+  card: {
+    backgroundColor: '#fff',
+    borderRadius: 12,
+    padding: 20,
+    marginBottom: 16,
+    shadowColor: '#000',
+    shadowOffset: { width: 0, height: 2 },
+    shadowOpacity: 0.1,
+    shadowRadius: 4,
+    elevation: 3,
+  },
+  cardTitle: {
+    fontSize: 20,
+    fontWeight: 'bold',
+    color: '#212529',
+    marginBottom: 12,
+  },
+  cardText: {
+    fontSize: 16,
+    color: '#495057',
+    lineHeight: 24,
+    marginBottom: 8,
+  },
+  boldText: {
+    fontWeight: 'bold',
+    color: '#007bff',
+  },
+  linkButton: {
+    backgroundColor: '#e3f2fd',
+    padding: 12,
+    borderRadius: 8,
+    marginBottom: 8,
+    borderLeftWidth: 4,
+    borderLeftColor: '#2196f3',
+  },
+  linkText: {
+    color: '#1976d2',
+    fontSize: 16,
+    fontWeight: '500',
+  },
+  footer: {
+    marginTop: 20,
+    paddingTop: 20,
+    borderTopWidth: 1,
+    borderTopColor: '#dee2e6',
+    alignItems: 'center',
+  },
+  footerText: {
+    fontSize: 14,
+    color: '#6c757d',
+    textAlign: 'center',
+  },
+});
